+++ app/mission/missions/internal_controller.py
@@ -1,3 +1,4 @@
+from app.mission.teams.model import UserTeamPositions
 import logging
 import os
 
@@ -48,6 +49,7 @@ class MissionInitPermissions(InternalAPIView):
         try:
             MissionService.init_mission_group(db_mission)
 
+            # admins group is organizer
             permission = DriveUtils.insert_permission(
                 db_mission.sd_root_folder_id,
                 "organizer",
@@ -56,6 +58,18 @@ class MissionInitPermissions(InternalAPIView):
             )
             if not permission:
                 raise SharedDriveException(KEY_SHARED_DRIVE_PERMISSION_EXCEPTION)
+
+            # mission creator is organizer
+            permission = DriveUtils.insert_permission(
+                db_mission.sd_root_folder_id,
+                "organizer",
+                "user",
+                db_mission.creator,
+            )
+            if not permission:
+                raise SharedDriveException(
+                    KEY_SHARED_DRIVE_PERMISSION_EXCEPTION
+                )
 
             permission = DriveUtils.insert_permission(
                 db_mission.sd_root_folder_id,